Enhance CycloneDDS and iceoryx integration for QNX and general improvements
- Updated CMakeLists.txt files to support QNX-specific paths for iceoryx and CycloneDDS installations.
- Added logic to handle shared memory (SHM) configurations based on the presence of iceoryx.
- Improved error handling in BufferCore for unknown error codes.
- Enhanced LoanedMessage class to support native CycloneDDS writer loans and fallback to heap allocation.
- Updated installation scripts for CycloneDDS to create configuration files for SHM and improved logging.
- Added support for optional ACL handling in iceoryx installation script.
- Enhanced test configurations to include iceoryx paths and libraries.
- Improved documentation and comments throughout the code for clarity.

+++ lwrcl/adaptive-autosar/tf2/src/buffer_core.cpp
@@ -665,7 +665,7 @@ void BufferCore::lookupTransformImpl(
       case tf2::TF2Error::TF2_LOOKUP_ERROR:
         throw LookupException(error_string);
       default:
-        printf("Unknown error code: %d\n", retval);
+        printf("Unknown error code: %d\n", static_cast<int>(retval));
         assert(0);
     }
   }
@@ -1459,7 +1459,7 @@ void BufferCore::_chainAsVector(
       case tf2::TF2Error::TF2_LOOKUP_ERROR:
         throw LookupException(error_string);
       default:
-        printf("Unknown error code: %d\n", retval);
+        printf("Unknown error code: %d\n", static_cast<int>(retval));
         assert(0);
     }
   }
@@ -1479,7 +1479,7 @@ void BufferCore::_chainAsVector(
         case tf2::TF2Error::TF2_LOOKUP_ERROR:
           throw LookupException(error_string);
         default:
-          printf("Unknown error code: %d\n", retval);
+          printf("Unknown error code: %d\n", static_cast<int>(retval));
           assert(0);
       }
     }
